pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ana-saga-cli"
version = "0.1.0"
description = "Agente de negociação avançada do SAGA, reconstruído para CLI terminal com arquitetura limpa."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["openai>=1.68.0", "PyYAML>=6.0.1"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
